(build) builtins for wayland-wm helper
in order to avoid more duplicated code (and eventually clean up the
atypes/wayland scripts in durden), the wayland and xwayland specific
window management rules have been added as a builtin/wayland.lua
script. See the included wltest interactive test for example use.

this also adds most of the string/table extensions that durden and other
projects rely on in their suppl.lua, so when this branch is merged
those can be dropped from related projects.