(core/shmif) allow hints in segment request
This is an optimization that should save a resize- synch delay in
mapping subsegments. A common pattern, especially for allocate-heavy
clients like wayland, is to set hints (partial updates, fd-signalling
and so on) immediately after acquiring a new segment.

This is likely to add up to a full frame delay before having access to the
segment. With this change, the segment request carries the desired hints
and will be applied to the subsegment immediately on allocation.

While this does not currently apply to the primary, it is possibly
worthwhile to have different standard hint-sets based on primary segment
type, especially when we default to server-side text rendering for tui
segments (as the tpack format toggle is a hints- enable action).

FossilOrigin-Name: ba97debee4230defb77fb482fc1a1642171c16fee0abe7f01462f53cbc0873d8