(builtin/mouse) several improvements to mouse.lua
This adds :

1. a possible fastpath that can be used to ignore multiple
dynamic handlers listening on the same vid. Set 'own_vid' when
registering the handler and it will have priority.

2. loading of keys from database for certain properties on init
these are prefixed with mouse_ as the keyname and should save appls
from reinventing the configure step for mouse tuning repeatedly.

3. the idea of a long-press gesture for left-mouse button, this is
both a safety mechanism for partially broken emulated mice (trackpads)
and as an accessibility feature. Long press of the button will remap
into a right click or a double-click.