(core/lua) let image_shader affect rendertarget
This one goes a bit deeper than it seems, but for dealing with
linktargets where a slice of one rendertarget is used for another,
(a decent primitive for certain effects) the basic shader use
might need to be overridden for the link target in particular.

image_shader is extended to handle a set of extra attributes where
a soft-default override shader can be defined for a rendertarget as
well as a force override.

With this another piece of ugly re-appeared that might have
contributed to the long-standing problem of certain terminal updates
taking suspiciously long to reach the screen but not when there are
animations pending. This suggests it is the dirty tracking that is
involved, possibly amplified in cases where the same feeds referenced
by multiple vstores. Rather than simply returning when a feed is
attempted to update multiple times between scanouts, we now flag the
rendertarget as having a pending transform.

More tuning to be done here, but not without first adding all relevant
paths to the tracing facility so we can get some proper data.

FossilOrigin-Name: 7bb7c795776133d29cfcf73580cd86ebd91c6ef1fdc752658748c65fe963bb92