(core/shmif/platform) more trace marks
The output (as shown in the tracetest) should now be sufficient to
get ok traces in chrome://tracing, though it is still a little weak
to take better advantage of something like tracy.

This also pulls in the trace collection into shmif, in order to
soon be able to collect data from the other side as well. The
implementation is only a few kilobytes and should be enabled later
through the builtin debugif.

Ultimately the arcan_tracer.c implementation should be swappable
for one that simply maps into tracy if such a build is provided,
then see what kind of mix would be valuable from the system_snapshot
perspective.

Regardless we are way past due for refactoring logging and tracing
to use the same paths and the same formats if the more complex perf
work is ever to progress.