(a12/lua/fsrv) rework net-fsrv integration
This adds the basic keystore to a12 and drops the last remnants of
the old net_ calls in the lua API. The (currently) one remaining is
a12_listen that now redirects to afsrv_net.

That is a first step towards then letting afsrv_net setup the a12
state machine for acting as an arcan-net -l that integrates seamlessly
with the lua layer.

The mapping / roles work as follows:

    afsrv_remoting (one main 'window', client provides input)
    afsrv_encode (other side to remoting)
    afsrv_net (letting networked clients connect, service-discovery)

the service discovery role won't be implemented until 0.6 series
as it requires key directory-server as well as broadcast- domain like
detection. It is also undecided whether define_recordtarget should
have a mode that act as outgoing client connection or if that is better
suited for extending define_arcantarget. The first can be fixed in a
few lines of code, the second takes more work but would also improven
the lwa 'app mode' semantics so likely going with that.

afsrv_remoting and afsrv_game are possible candidates for deprecation
or, rather, being moved into stand-alone tools as their dedicated roles
did not pan out as important as other mechanisms evolved.