(core/lua) forceupdate adjustement, rtgt- framecookie
This addresses a possible bug where rendertargets that are updated
out of bounds and at the same time mapped to a display won't be drawn.

The edge case is that a linktarget might cause worldid- to be updated,
so that when the normal refresh arrives, it no longer has any 'dirty'
objects.

Eventually the entire 'n-dirty' mechanism should go the way of the
dodo, and this rendertarget tactic is somewhat better - though there
are still some weird edge cases with partial updates and partial
scanout (where that is supported), n-buffered rendertargets and
same target mapped to multiple displays (can happen with different
coordinate sets).

Forceupdate itself also gets a boolean flag format where it is
possible to force an update, but only if the rendertarget is actually
dirty. This is useful to rate limit certain advanced effects.

FossilOrigin-Name: 659d694df43f129e131f8f4b37bc673fb97c30feee1316212655eb9b54cdbd01