(core/tui/lua) bchunk type hints rework
This allows each tui window to be able to accept bchunk out store
requests in a raw format, capturing the front buffer as (for now)
unformatted utf-8. More formats can be added later to also get a
markdown or html friendly output immediately, but baby steps.

While doing so some ergonomics in how bchunk hints were delivered
needed to be done to make it easier for scripts to harmonize the
'accept'/'provide' list of types for drag and drops of multiple
types, something that was added to be able to deal better with how
the data device works in wayland.

FossilOrigin-Name: 96c5bab3249a334a1b9dd982bcdd8c90933869135dc0f1f8edcaf0eb9ea3c717