(build) temporarily disable -Wdeprecation
The annoying ffmpeg encode deprecation warnings had got to go as they
got in the way of absolutely everything.

In true ffmpeg fashion, the changes to the api are basically such that
we need to rewrite almost everything, and that won't happen this release
window. It will likely be a while longer until they can touch the deprecated
functions (heck initially they were marked deprecated even before the
alternative was there, provide factory wrappers already for god's sake).

Nothing else should be vulnerable to deprecation, so suck this up for
the time being and next time around rework encode to keep ffmpeg as an
optional dependency and do some damage control.