(core/tui) tpack tracking cell dimensions
This should fix one of the few remaining core issues with tpack for
server side rendering. There are some test-cases more to run when
it comes to hot swaping fonts, dynamic changing density as well as
differences in cell size probing -- otherwise *fingers crossed*
the worst part of this one is over and the feature can soon be set
to the default (and thereafter, the only option).

Mental checklist for later versions:

 * shared glyph cache between frameservers
 * glyph-cache to atlas
 * tpack-renderer that slices from atlas
 * option for MSDF atlas (offline / background gen?)
 * image_textsurface (for TPACK access from lua)
 * update image_access_storage to allow glyph access
 * re-add shaped lines (with feedback)
 * re-add substitution (with feedback)
 * re-add smooth scrolling