(fsrv/lua) let clients control input masking
This adds the option for clients to add INPUT_MASK events signifying
device and/or data types that it does not support. If the appl tries
to input such an event, it will be silently dropped rather than added
to the queue. The 'input_capabilities' function has been extended to
query the current stage of such a mask and the event handler has
a 'input_mask' event added to signify that the capabilities have
changed. This allows appls to provide emulation/remapping for a
device type when it is not supported.

FossilOrigin-Name: d433acc3f77e2d23219ce43a8c1a75c555bb6ebeb214a30a05a63f06cf1f6eeb