(lwa/lua) flesh out def_arcantgt/arcantgt_hint
This has been pending for a long time, and while though there still
are some bits left to it, the feature is usable enough for experiments.

Basically the 'lwa' mode previous did not act as a full shmif client
as the event processing was missing, and registering subsegments to
usable windows was only partially implemented.

This rather big patch adds an 'arcantarget_hint' function that allows
sending hints such as labelhints (like normal clients would be able to)
as well as a change to the (still experimental) define_arcantarget
function that removes readback/update controls in order for the server
side to control the clocking better.

Some things are missing still, the most glaring one being that subsegs
still don't transfer as the WORLDID rendertarget does (default dma-buf
with software reading fallback) but only as a readback. Part of the
reason is that we want to be sure that the modifier and fencing
structure will be the right one at first go, which is for a later
release.

The second is that just a small set of the events/commands are filled
out, but that is a less complicated procedure and can be handled on a
'add a command, update doc/test, commit, repeat' basis.