(lua/core) add scale method to link_image
This one should make it much easier to make attached objects that
follow the scale and transformation of its parents. Previously if
something like an icon or border would link to a parent and be set
to resize at the same pacing as the parent, rounding errors and
different stages on the curve (worse with non-linear interp) would
cause glaring visuals.

The fix is to allow the child to specify which dimensions it wants
anchored to the parent:

    link_image(a, b, ANCHOR_LR, ANCHOR_SCALE_W)
    resize_image(b, 100, 20, 1000)

this will now cause 'a' to maintain the same width as b.w + a.w while
a.h remains a.h. Either or both dimensions can be linked this way.