(lua) new entry-point: _input_end
In order to better support batching of input events and accumulation
of analog axes during event storms there should be some indication
when event processing has completed.

Previously the _preframe_pulse could be use as such a trigger, with
the caveat that it won't work when we can do (limited) event processing
during scanout.

Therefore this adds a new _end entry point that will be called if
events have been flushed from the queues. Note that this covers _any_
events through the queues and not just input events.