(lua) add relink_image
This adds a special variant to link_image that takes the common pattern
of wanting the world-space coordinates of the src remain the same when
translated into the object-space of its linked parent.

Doing this in script alone is somewhat tedious and costly, and makes
more sense as a separate function.

FossilOrigin-Name: ceff5211c37320fd84298cfaa3edef69201e4f9af699cd17d001e353b2de1378