(tui/arcterm) add =pipe mode and stdin/stdout map
This adds a bchunk handler inside tui for taking 'stdin', 'stdout'
and 'stderr' as types and duping into the corresponding descriptors.

It also adds a 'pipe' argument where the normal pty is setup for
running the client and feeding the state machine, but also mirrors
the output to a pipe.

This allows WMs to dynamically splice and create multiple views of
the terminal input/output.