(core/lua) blend-modes and blend-default
This tries to correct a long-standing wart (one of the oldest ones)
by first making the different blend modes actually do something as well
as adding some missing ones. There are more combinations in the
traditional GL model of course, but for more exotic variants the tactic
is still to build your own using rendertargets.

The larger issue though is that the behavior for the default
BLEND_NORMAL was wrong and actually behaved like BLEND_FORCE. To
mitigate this we add a control for switching the global default blend
mode (text objects excluded) and sets the default to BLEND_FORCE as the
proper BLEND_NORMAL will break quite a few applications.