(lua) identifier argument form for nonblock/bond
With bond_target it was previously possible to indicate the type used
for the BCHUNK_ hint. The same was not possible when a fsrv was
referenced with open_nonblock, requiring the receiving end to always
handle 'stream' or '*'.

FossilOrigin-Name: 038a7ef77009fa811baa95035c3197607c6e80e46a6cd566e185e66485c17183