(net/shmif-srv) more granular shutdown
This should fix the "fun" bug of arcan-net -s killing clients early when
run in forked mode. The problem was that the cleanup shutdowned the
socket even in 'no_dms' mode. The fix is to add another destroy call
that do preserve the connection but cleans up the resources in the
parent process space.