(decode) basic uvc support mostly working
This add some format and dimension selection to the uvc capture mode.
NV12 is dealt with by putting throw swscale (as should the other fmts
really), as we need it later for MJPEG and H264 as well.

Other things to expose would be specific device control, activation
handshakes for various exotic devices (LeapMotion, Oculus cameras and
the likes).