(decode) use swscale for yuyv2, ...
This is (very) likely not correct, though I don't have enough UVC
devices to test for real - more UVC work will be needed later anyhow to
map MJPEG/H264 as well as exposure controls and so on. There looks to
be some UVC bugs that should be fixed upstream -

1. not getting the format requested (even though we ask for NV12 it
   will pick the first with matching dimensions).

2. buffer allocators need to be outsourced so we can guarantee alignment
   that fits swscale as well as map into the PBO buffer directly to get
   faster GPU streaming.

Other cameras tested (LeapMotion) seem to require secret handshakes to
actually initiate capture, so that is another blocker to deal with.