(agp/lua) add viewport/scissor rtgt control
This adds arguments to resize_image_storage to define the coordinate
range that the rendertarget will process. The main idea for this part
is to allow a define_linktarget to process a different part of the
same pipeline as another rendertarget.

FossilOrigin-Name: 68f28f80576d18d730d8043a532b559429cbd304bbf4cc29eeab97111e93f1c2