(core) reset cell dimensions on recalc.
This caused a subtile issue when going from a larger cell size to
a smaller one for tpack rendering. This could happen when manually
switching to a lower font size or reassigning to a rendertarget
with a different density. The cell size probing was based on having
a list and repeatedly calling the probe, hence w/h acting as state
between calls. This was not reset when re-probing, meaning that the
larger cell size would be retained.

FossilOrigin-Name: 45dafa797f1a528829744a486cd8a1f76eaa13650a752baad1e6f4969361b5f4