(agp) don't blend when drawing to primary-rtgt
This caused some issues on SDL2- like platforms where (currently) the world is
drawn to a rendertarget that is then re-blit into the 'screen' itself. While
the reblit is undesired generally it is necessary at the moment to harmonize
with native platforms where we can export/import FBOs. Anyhow, when performing
the reblit stage, blending was erroneously added and if the rendertarget being
reblit preserved alpha you'd possible get a shadow of the previous frame in
some cases.