(core/shmif) patch in modifiers/multi-plane
This adds a long-pending patch for getting in multiplanar and
modifiers into the gbm- based platform, which was one blocker for
defaulting arcan-wayland to use dma-buf over legacy drm.

As a bonus, the frameserver code is refactored to better handle
cancellation and failures in mail-box mode for clients, and there
is less needed for explicit synch when those kernel patches eventually
come.

It still has subtle and not so subtle breaks on CoffeLake GPUs at the
least, using the tests/frameservers/glcube will assert/fail on debug
MESA builds, and give graphics glitches otherwise. In order to get
further with that we need better allocator control for rendertargets
with 'swap' capabilities, so that should be ironed out (which is a big
performance boost anyhow so).