(tui/terminal) autofit to keep_alive, contenthint
This changes the tui_wndhint to forward both a content hint and a
viewport event depending on the argument in order to communicate
the min/max cell-span of the content.

The terminal gets an autofit argument to keep_alive that will try
and figure out the useful boundary when a client has died. This also
tries to maintain screen contents on window resize after pty- death.