(tui) cherry-pick i8n/havoc lost-lines on resize
This covers an ugly edge condition in tsm_screen that would case resize
events in normal-screen mode to needlessly scroll contents of screen
even when it would fit. Since it will be a while longer until
tui-linewnd allows us to remove tsm screen entirely, this bandaid will
have to do.