(wayland) configure with cont- size if !set geom
This caused mpv and friends to jump back and forth in size on selection
state changes.

The interpretation of the spec here is liberal (orly) - so for configure
it is supposed to be the 'geometry' (e.g. boundary of subsurfaces
including decorations), and if we don't know (at setup) then 0,0 and the
client is supposed to decide.

The client does and sends it, but before it arrives we can get a hint
about what the new size is supposed to be from up high. So we send the
new configure, but then the client comes with its geometry and new
buffer commit and so on.

Longer story short - if geom is not set on a hint, use the current
context one, if the problem persists, switch to the last hinted one.