(agp/shmif) buffer-allocator interface progress
Should be about 2/3rds of the way for this one now. It is likely that
some of the glitches we have on certain Iris+ intel GPUs are tied to the
fact that FBOs in arcan_lwa are not allocated with a color store that
carries format/modifier which allows it to be shared as a dma_buf.

This is similar to the problem we have with not doing direct fbo scanout
in normal arcan/egl-dri, which is a big performance hog. The reason for
this detour was to be able to have a fbo->shader->screen pass from the
start, but you should also not need to pay the price unless necessary.

This required the agp- layer to be re-written so that FBOs have a
swappable allocator function that tracks formats based on outgoing
display and so on.

This patch tests that agp/glshared does use the allocator interface
correctly (ish, deallocation misses a spot on resize), as well as
plugging in the BO-EGLImage-GLid allocation path needed in shmifext.

Left is to:

1. forward modifiers correctly via DEVICEHINT.
2. plug them into the allocator, and switch to it.
3. allow a similar path in egl-dri when sane_direct_vobj applies.
4. add the allocator- provided BOs as framebuffers.
5. do the same for client allocated USE_SCANOUT flagged buffers.
6. handle the ensuing resize- edge case mess.