(egl-dri) improve edid caching
When a display was removed, the previously known edid was also cleared.
This is to avoid assuming that the next display in the same connector is
the very same display.

At the same time re-querying the edid might fail. In that case any
future edid requests from the platform layer would always fail. This
patch will attempt to re-query the EDID if it is still missing when the
WM asks for it.