(...) tpack-dumb cross-blit working
There are some caveat and subtle races there still during handover
that seem to be able to cause gbm to stop double buffering.

This config makes sure that if there is a local vstore, the blit region
for tpack will be cloned there. That, when combined with re-setting the
Crtc on each dumb buffer update (wonder why this is needed for reliable
update) and the dedicated-fullscreen terminal case is actually working.

The real savings still require the asynch-input forward path, proper
vsync updates on client, glyph cache sharing and so on, but this is more
hinged on the rest of the text rendering updates as well along with
cursor overlays.

FossilOrigin-Name: 4a1f4ff99372813c925894ece89c049d27723934ead3c70e56a506f249742cc1