(core/platform) layer-mapping and vstore-dirty
This continues the single-buffered output path for egl-dri and friends
with the ability to signal which parts of the surface that is dirty.
Some friendly gpu drivers will leverage this to only synch-invalidate
caches for affected areas.

Since it involved adding functions to the platform definition, another
related change is added as well, platform_video_map_display_layer.

This takes the place of platform_video_map_display as it allows mapping
multiple objects to different output layers, if the hardware supports
non-gl composition. This still requires changes on the Lua level, and
the first likely candidate is for accelerated mouse cursors and similar
overlay- type surfaces.