(egl-dri) fbo-scanout should not swap on !dirty
This brings (primary) screen fbo-bo scanout path working on primary.
There is something else dangling for secondary screens, but babysteps.
The necessary change was simply that the update-direct return path
indicated that we should flip even on a rendertarget that wasn't
damaged, causing the previous frame to flicker on and off.