(platform/lua) add options to video_displaymodes
This is the next step for VRR and HDR handling. In addition to the
normal display modesetting, it is now possible to provide a table of
options as to the targeted refresh-rate and desired quality level on a
display.

FossilOrigin-Name: 7b1140d7b220f0342bab58daed8c4ef79e6c8ba149f0ea62bf6d7eba545670b8