(lua) let displayhint and fonthint return cellw/h
This is an interesting sort of edge case to tpack clients (so tui)
during the preroll stage where the dimensions are supposed to be
expressed in pixels, which is dependent on displayhint and fonthint.

While the client can forward its view through contenthint, that is too
late for the initial buffer allocation, and since we want zero resizes
if at all possible, the current state should be queryable, and the
results of a hint should be predictable.

To support that case, these two hint functions will also return the
current estimate from the tpack- state in the frameserver tracking
struct.

FossilOrigin-Name: fdd5364bc9c780a411e854cbdce97d34d57e6581cfb05dd93be13a39ff46e2b9