(shmif) preroll-stdout dup failing from bad test
In the preroll stage, it is possible to override stdin/stdout from the
server side. This is mainly used for some specialized tools with the
terminal emulator. The stdout- part turned out to be broken as STDOUT
was not a permitted dup target.

FossilOrigin-Name: 673259159a6e4fef0108713e7dbb5525431a30b9fdaa8797f4236ce9f711c038