(terminal) pipe mode buffering fix
The previous approach simply tried to map stdout/stdin and project over
pty. This is flawed for the normal 'pty is actually serial and buffering
is garbage' reasons and it can randomly stall on top of adding various
TERMCAP noise.

This change modifies shl-pty to simply dup-in pipe-pairs and copy-in-out
from stdin/stdout through that. Some options are missing for this,
mainly a counter mode (like pv would since we are paying the price),
bufferwnd-edit mode, cleaned up non-state-machine presentation of
viewable range and some kind of direct-mode where the current stdin/out
is just forwarded into the client.

The direct mode is problematic in the sense of the meta-data to present.
It will clearly be faster since a full copy / rebuffer pass is lost, but
at the same time there is not that much interesting information to be
queried from the resource itself, and the client is given free reign of
the descriptors. The best option is probably liveness + stderr contents.

FossilOrigin-Name: 843f78f1dca5d2abf34db32f959b3b5440f5ae0a5fc181d9363b81db8c89a95e