(sdl2) utf8 missing in input workaround
This is yet another band-aid on the unstichable wound that is text
input in SDL2. For the sake of record, the core matter is that the
keypresses used to carry some unicode codepoint estimation in 1.21
days. This did not really work well with languages requiring IMEs,
but for the rest it was OK.

With 2.0 the TEXTINPUT event was supposed to provide higher level
text instead, BUT correlating that to a specific keypress instead
became impossible. With durden et al. we need both or provide our
own IMEs and overwrite the UTF8 translation...

The workaround was to provide some rough translation when possible
to make the burden of remapping at higher levels more tolerable.

The bug was that modifier part of remapping failed to acknowledge
latched modifiers like scroll/num/caps, emptying out otherwise ok
utf8 translation.