(tui/arcterm) invalidate screen on recolor
This should fix the last missing piece for the dynamic colorscheme
switching end. The restore-buffer stage will erase before applying
so even if non-indexed color default attribute is used, the change
to default-attr will be reflected.

Similary, when on_recolor has been triggered, mark the entire scrn
as dirty so that attribute-comparisons won't early out because the
>index< remain the same but the resolved color is different.