(tui/arcterm) more colors to indexed
The widgets rely less on defcattr now.

Any implicit / default attribute functions are really things that should
be delegated to the past and removed outright. They cause more problems
now due to the possibility of color-indexed attributes, and there is
still time to fix consumers of the API without too many tears.