(shmif) shutdown on EAGAIN in preroll
In some conditions, the blocking recvmsg could still be awaken due
to signal delivery and so on. If that happened during preroll,
chances were that it was interpreted as an unrecoverable error with
the socket, causing a shutdown.

The likely reason this hadn't occured before is that very little
data exchange has been performed in preroll. With recent changes to
allowing color scheme transfers in this stage, the chances for this
bug to hit increased.

FossilOrigin-Name: b85633788fffb325a23acbeb99a08088e17f23f98df47aa79796f7c324aa2840