(shmif/fsrv) add last_words on cookie-fail
While the shmpage itself has protections from partial corruption
with validation cookies at strategic locations, when those tripped
there was no communication onwards that this was the reason.

This patch adds a last_word (message to 'terminated' event) that
shows if the validation stage was triggered, as that is a strong
indicator of bad client code or something more foul.