(shmif) fallback to $HOME on failed $XDG_RUNTIME_DIR
Previously we used the priority order of XDG_RUNTIME_DIR and, if not
set, fallback to dotfile in homedir (though we should also consider
the scratchstore for arcan namespaces, but not yet).

In some exotic cases, we might be in a system where XDG_RUNTIME
isn't mandated yet some arcan instances might use it, and some not.

To handle the edge cases of recovery and switch between the two,
first try XDG_RUNTIME_DIR and if present but failing, switch to
a fallback before giving up.