(lwa) change bufferpass- fallback
The BUFFER_FAIL event is no longer forwarded, and the client is expected
to check if passing is disabled rather than to try and do something with
it itself. This failed for lwa, so fixed the check.

This should make bizarre chains, e.g. arcan_sdl -> durden -> lwa work
again, as well as lwa over a12. A related note is that SDL2 backend
always rejects buffer import still, which makes sense for when the SDL
backend is used in X11/OSX as we don't get the proper EGL context to
work with, but less so for Wayland. Since the drm dance is abstracted
internally now, we can probably just import/re-use that code when the
proper extensions are detected.

FossilOrigin-Name: 226104fc4eda36042d1128366e900f87b3a54c4e857587aab83c0647a9c8a3f5