(lua) tighten launch_target
The documentation and implementation for this one was grossly incorrect
when it came to overloaded function forms. The reason for this is an
earlier omission where a git commit -a added the documentation to
discarded changes.

Fortunately, nothing relied on the extended form that permitted an
argument to be appended - and with recent changes to afsrv_terminal
alternate CLI modes it is not needed. The cli should be used to
chainload any custom-external binaries.

What is still missing from this form of launch_target is to permit the
LWA launchmode to spawn new arcan_lwa instances now that this binary
is a default build. To make this slightly less cumbersome, the applpath
that is valid for system_collapse should also be valid for
launch_target.

FossilOrigin-Name: 5b8bf3acd28683ed8ad9c9c7894bd404f52b954285fa33b02bde3d62bcba1704