(wayland) subsurface on buffer-less surface hack
So some special clients are looking for ways to make subsurfaces worse
and suceeding. It seems to be 'tolerated' behavior to have subsurfaces
attach to parents that themselves have null-buffers.

This patch works around it to the extent that the clients won't crash,
but not render correctly either. This happens with firefox. If that
matters, use the x11 version or switch to something less awful, like
qutebrowser.

FossilOrigin-Name: 91bdf5e8d43252762d2ab797fad8af5285034a4eba33b4911c581755c4520a53