(core) suspend- audio device with platform
The audio layer is still severly neglected, but an easy openAL win
is to suspend the device mixing when the platform is not active.

This is, as with many other things in AL, an extension that is not
supported across the board so have to be probed and treated
accordingly.

The implementation here is not entirely correct as the lookup funcs
are per-device context, though we only support the one at the moment
so.

Other improvements sorely needed here (on top of a platform split
so that we are no-longer dependent on OpenAL-soft and its LGPL taint)
is to also apply this heuristically when there are no active sources
that are likely to provide audio.

A good test case is arcan_lwa which will burn a lot of juice mixing
and forwarding silence, when the processing could just be suspended.