(shmif) slightly less lenient guard-thread
This continues to work poorly for all possible combinations, error
points and modes. If this still yields 100%- spinning clients in some
corner cases.

The reason for this idiocy in the first place was some kind of
compromise that mostly worked on the more bizarre platforms (OSX, Win32)
or, at least, with the least amount of per-platform code. Since those
are slowly deprecating themselves, we can reconsider.

The next step then is to abandon this broken idea and simply push a
DEVICE_NODE event with an integrity pipe FD acting as the keep-alive
channel for the guard thread to monitor.

This should work both for the authoritative and non-authoritative modes,
and DMS can be left as an in-process release mechanism. Since the V&V
is non-trivial, this and the 'delay event slots' should hopefully
survive being band-aided until shmif is refactored into a platform/...
and libarcan-abc is realized.