(shmifext-gbm) _dev render-node extraction
A subtle and complex one, but some rare edge cases (xarcan passing
clients along being one) impose that a render-node being opened isn't
actually allowed to be DUPed from an existing one without consequence
but rather has to be opened uniquely.

This patch recalls the last known DEVICE_NODE path and provides that
when a client requests a reference to the device node. This will not
work for GPU redirect where a new device node is passed, as that takes
further changes.