(a12) raw vframe handler was never called
This should only matter for test applications that use the hook path
instead of through a shmif context. In hindsight, this was a mistake
from the start, and the real grunt work of having a valid/hollow in-
memory shmif implementation should have been used instead for better
code re-use.