(a12) helper: don't want for more bytes if preauth
This caused a stall and a possible discarded first message if the
authentication was set to skip x25519 entirely and just use pre-shared
symmetric keys. In practice this wasn't noticed due to how loud durden
and friends are on startup, so there would always be data queued on the
line.