(a12) encode:raw overflow on chunk-padding
(untested) if the method is forced to the (mainly for testing) raw
formats, and the source dirty region of the buffer size did not align
with the rough 'frame chunk interleaving window size', the wrong number
of pixels were being added to the short chunk (the code looks to have
been just copy and pasted from above without actually accounting for
the proper values).

If this works, the same procedure should be used for 565/rgba as those
are copy-patch-pasted as well.

FossilOrigin-Name: 357942a4c40ff04b8dbda103c1d55333a090fb19a79450279219384b3dfdccc2