(egl-dri) cache modes on discovery
This is a pattern that has been applied by some WMs in the past, notably
durden. It makes little sense on that level though and should just be
added here. The problem/cause is that scanning the display for possible
modes can have a noticeable (hundreds of ms) stall, while the actual
state should only change on hotplug events.

This patch simply ties the mode lookup to the display upon
hotplug/hotunplug and returns that cache when available.