(tui) add -DTUI_RASTER_NO_TTF build flag
This removes the freetype dependency from all tui clients, and this
can eventually be switched to the default. The caveat is for shaping,
ligatures and emoji/line-drawing characters as the client side is not
aware of which glyphs are actually present on the other end.

Extracting the tables themselves and having a smaller parser for the job
is certainly an option - albeit one that takes quite a lot of work.