(a12) pull in zstd implementation
In source for embedded/reliability - while reasonably well packaged
across distributions, ensuring that the a12 implementation has the same
view on build types and so on is a bit too important for extdep.

This does not add the integration, when that is done miniz/zlib will
be evicted entirely* (still support decompress for a while to not break
current interop, as well as to allow for the distant option of
passthrough transfer of PNG sources).

FossilOrigin-Name: 243f92ade27848a57ab687c9f60b968991b344b45cb1cefcf98cdad8eabd546d