(a12) deflate to zstd working for image
TPACK is still going through deflate, but shouldn't be more of a
procedure there. The missing items still are that there isn't a proper
context cleanup on channel shutdown, and the same applies afact for
H264.

Otherwise the next step is to actually calculate the congestion window
and start delaying producer release and increase compression level when
window is full.