(a12) pingback on frame completion
This is the first step in adding better congestion control. When a batch
unpack is completed, a return event is queued so that the other end gets
its 'last_seen_seqnr' updated. The difference between the seqnr when a
frame is sent on a channel helps determine if too much backpressure is
accumulating and the sending side should back-off by deferring source
'release' and/or increasing compression level.