(egl-dri) revert default-direct + lua workaround
There are still too many subtle edge cases with the direct rendertarget
scanout, either through bugs in our code, drivers or, worse, both. While
safe 'most of the time', it is especially troublesome with resolution
switches, display hotplug and display rotation where we can get 'ghost'
frames coming back to haunt us.

The tactic now is to revert to 'default compose' and change the config
option so you have to add:

    arcan_db add_appl_kv arcan video_device_direct_scanout 1

To get the faster but somewhat unsafe behavior. In order to test this
more easily, a flag has been added to map_video_display (HINT_DIRECT)
that will do this on a case by case basis dynamically. While this is not
supposed to be a long term solution, this can be invalidated later by
just making the flag resolve to 0 (HINT_NONE).