(lua) include worldid in snapshot rtgts
Snapshots created through system_snapshot (actively or as part of script
crash collection) did cover secondary rendertargets but the primary one
was not included.

That made it unreliable to reconstruct the scene graph traversal order,
as the cellid+order+inherit_order flag is not enough for that when it
comes to objects with the same order value.