(lua) reduce number of strlen calls
This fixes maybe 70% of the unnecessary calls to strlen when moving
data in and out of the VM. The problem was the use of pushstring when
key is a literal or when both key and val is a literal. What remains is
the cases where we do an explicit filter-copy and want to re-use that
(look for the fltpush/MSGBUF calls). For those the length should at
least be propagated and switch to pushlstring.

The way this is worked around is quite dangerous, as any dynamic
argument mistreated through tblstr will fail silent with truncation to
pointer size - hence why it is staged based on patterns.