(tui) pre-dup bchunk file-descriptor
This behavior was left unspecified while well-defined in shmif,
and that is that the descriptor is only alive for the span of the
callback and has to be dup:ed manually.

For shmif this makes sense as there is no indicator that the
client will react to the event and consume the descriptor outside
complex cases e.g. handover-exec and newsegment.

Thus for tui contexts, if bchunk is set as a handler - the client
is expected to close the descriptor, and will only be invoked ifg
an internal dup works.