(tui) pre-dup bchunk file-descriptor
This behavior was left unspecified while well-defined in shmif,
and that is that the descriptor is only alive for the span of the
callback and has to be dup:ed manually.

For shmif this makes sense as there is no indicator that the
client will react to the event and consume the descriptor outside
complex cases e.g. handover-exec and newsegment.

Thus for tui contexts, if bchunk is set as a handler - the client
is expected to close the descriptor, and will only be invoked ifg
an internal dup works.

FossilOrigin-Name: 9c6048850451192a84f772037f825b605a3786f5ff44d19e9bad7bdcd2f903f3